scaffold: initial position param
It's hard to build complicated scroll UI, such as:
1. List view with a dynamically sized footer = height of other widget
2. That starts at the bottom when rendered, like a chat.

#2 is possible by reversing the list, but, it causes other downstream problems. IIRC it also didn't work well with #1.

Client apps *can* just trigger an immediate scroll to bottom, but, this results in rendering cost during the worst time: say, a chat comes on screen. Now it has to render the top + cache area, then it gets told to scroll to bottom, and has to render that too.

This is a trial of offering a param that tells the list where to start. It's working, but, there's still an issue.

At least, in the complicated case of Telosnex, where it is rendering large chat messages into Markdown.

The scroll might not *actually* scroll all the way to the bottom. Close, but, not there.

Checking it in for now.